+++ noodles-sam/CHANGELOG.md
@@ -20,6 +20,8 @@
 
 ### Changed
 
+  * sam/async/reader: Read record into a `Record` buffer.
+
   * sam/reader: Read record into a `Record` buffer.
 
   * sam/record/cigar: Change conversion from `Vec<Op>` to be